requirements.txt
requests
streamlit
openai
python-dotenv
beautifulsoup4
lxml